challenge1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
plotly==5.17.0
